UI updates from product feedback
